// src/app/models/modules-enum.ts
export enum Modules{
  Abruf = "Cannot Delete Abruf",
  PO = "Cannot Delete PO",
  PG = "Cannot Delete PG",
  Consultant = "Consultant",
  ClientRemove = "Cannot Remove Client from the Project",
  removeConsultant = "Cannot Remove Consultant from PG",
  noEndDateConsultant = "No end date in consultant",
  Project = "Cannot Delete Project",
  Client = "Cannot Delete Client",
  Timesheet = "Cannot Enter Timesheet Entries",
}