// src/app/models/po-pg-consultant-obj.ts
export interface ObjPoPgConsultant {
    poNo?: string;
    poId?: string;
    poStartDate?: string;
    poEndDate?: string;
    pgId?: string;
    pgName?: string;
    consultantId: string;
    consultantName: string;
    consulantStartDate: string;
    consulantEndDate: string;
    assignmentId: string;
    flag: boolean;
    poTimesheetTemplate?: any;
  }